Reorganize handling of high-level TEI elements
The stylesheet is now structured to identify the root element of
the TEI file. This will allow processing of <teiCorpus> later.
For MODS elements which are the same for every TEI file, the XSLT
generating those elements has been moved to the template handling
<TEI>.

Abstracts are generated from <abstract> or <div type="abstract">,
and in a modular fashion (using <xsl:apply-templates> rather than
<xsl:if> or <xsl:for-each).