Streamline names handling
Names are now separated into <mods:namePart>s with @types of
"family" and "given" where appropriate. If the TEI-encoded names
include @sort, then a <mods:displayForm> is populated using the
specified order.

Also, all names that are children of TEI's model.respLike are
treated as separate entities, so, for example, a <tei:author> may
contain the names of one or more personages.